Move File{PutChar/GetChar/Printf} to libmisc.
